llama-openrouter-demo/requirements.txt
llama-index>=0.10.0
llama-index-llms-openrouter>=0.1.0
llama-index-embeddings-openai>=0.1.0  # For vector embeddings in advanced example
python-dotenv>=1.0.0
aiohttp>=3.8.0  # For async HTTP requests in streaming example